@@ -4,7 +4,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -14,6 +14,9 @@ dependencies:
   cupertino_icons: ^0.1.2
   flutter_bloc: ^2.1.1
   meta: ^1.1.8
+  http: ^0.12.0+3
+  shared_preferences: ^0.5.6
+  string_validator: ^0.1.3
 
 dev_dependencies:
   flutter_test:
