@@ -5,15 +5,29 @@ abstract class LoginEvent {
 }
 
 class LoginButtonPressed extends LoginEvent {
-  final String username;
+  final String email;
   final String password;
 
   const LoginButtonPressed({
-    @required this.username,
+    @required this.email,
     @required this.password,
   });
 
   @override
   String toString() =>
-      'LoginButtonPressed { username: $username, password: $password }';
+      'LoginButtonPressed { username: $email, password: $password }';
+}
+
+class SignupButtonPressed extends LoginEvent {
+  final String email;
+  final String password;
+
+  const SignupButtonPressed({
+    @required this.email,
+    @required this.password,
+  });
+
+  @override
+  String toString() =>
+      'SignupButtonPressed { username: $email, password: $password }';
 }
